Add files via upload
Alter the code of this file to create other size tables of random QR codes with derived colored borders. If you scan in and append four of these together, then they make a n EVM public key. The scanning tool that I use isn't included at this point. 

This is covered by a lesser GPL license, so you can use it for commercial stuff. I informally presented this at Eth Denver 2024 and perhaps should have put it into the hackathon. Don't claim that this is yours and sue me. This is less than 100 lines of code and would function best as something that anyone could print off and use. How would you even try figure out who as run it in a javascript session and then printed it off locally?

Anyhow. Have fun.